+++ source/EngineGpuKernels/Object.cuh
@@ -7,10 +7,10 @@
 
 struct CellMetadataDescription
 {
-    uint64_t nameSize;
+    int nameSize;
     uint8_t* name;
 
-    uint64_t descriptionSize;
+    int descriptionSize;
     uint8_t* description;
 };
 
@@ -49,14 +49,14 @@ struct ConstructorFunction
     int constructionActivationTime;
 
     //genome
-    uint64_t genomeSize;
+    int genomeSize;
     uint8_t* genome;
     int genomeGeneration;
     float constructionAngle1;
     float constructionAngle2;
 
     //process data
-    uint64_t genomeReadPosition;
+    int genomeReadPosition;
     int offspringCreatureId;    //will be filled when self-replication starts
     int offspringMutationId;
 
@@ -93,7 +93,7 @@ struct InjectorFunction
 {
     InjectorMode mode;
     int counter;
-    uint64_t genomeSize;
+    int genomeSize;
     uint8_t* genome;
     int genomeGeneration;
 };
